Connected Campaigns from MAUI interface to API
Added new properties to the Campaign model such as ImageURL and AllocatedBudget.
Changed the Name field to Title.
Copied the Campaign model over to the MAUI project until models are migrated to a class library.

Created a CampaignService with its associated interface to handle API calls.
Registered the service in the main program.

Prevented the API from automatically launching a new web browser page.

Changed the MyCampaigns page to load data from the API using the service and added actions for buttons.

Changed the route the campaigns and reflected the change in the navigation bar.

Registered the CampaignService in the API as well.